split components in separate files
